refactor(saga): Phase 5 extract ISagaTransitionRunner + EF impls; processors delegate
- ISagaTransitionRunner<TState,TEvent> + TransitionResult<TState> in
  Domain/Abstractions/. EfOrderSagaTransitionRunner adds saga-specific
  BeginCompensation(sagaId, origin, trigger, ct) overload.
- OrderSagaReplyProcessor + RefundSagaReplyProcessor reduced to one-line
  RunAsync delegate; SagaReaperService reroutes compensation through
  runner.BeginCompensation rather than poking SagaContext directly.
- Old OrderSagaTransitionResult/RefundSagaTransitionResult deleted;
  state machines + OperatorSagaEndpoints + handlers now consume the
  generic TransitionResult<TState> (field shape unchanged).
