--- saga-microservice/Saga.Service/Domain/OrderSaga/OrderSagaTransitionResult.cs
@@ -1,8 +1,0 @@
-using ECommerce.Shared.Infrastructure.EventBus;
-
-namespace Saga.Service.Domain.OrderSaga;
-
-internal sealed record OrderSagaTransitionResult(
-    OrderSagaStateSnapshot State,
-    IReadOnlyList<Event> Commands,
-    bool Changed);
