--- saga-microservice/Saga.Service/Domain/RefundSaga/RefundSagaTransitionResult.cs
@@ -1,8 +1,0 @@
-using ECommerce.Shared.Infrastructure.EventBus;
-
-namespace Saga.Service.Domain.RefundSaga;
-
-internal sealed record RefundSagaTransitionResult(
-    RefundSagaStateSnapshot State,
-    IReadOnlyList<Event> Commands,
-    bool Changed);
