+++ saga-microservice/Saga.Service/Infrastructure/Reaper/SagaReaperService.cs
@@ -4,6 +4,7 @@ using Microsoft.Extensions.Options;
 using Saga.Service.Domain;
 using Saga.Service.Domain.Abstractions;
 using Saga.Service.Domain.OrderSaga;
+using Saga.Service.Infrastructure.Data.EntityFramework;
 using Saga.Service.Infrastructure.Observability;
 
 namespace Saga.Service.Infrastructure.Reaper;
@@ -50,11 +51,13 @@ internal sealed partial class SagaReaperService : BackgroundService
 
         var sagaStore = serviceScope.ServiceProvider.GetRequiredService<ISagaInstanceStore>();
         var outboxStore = serviceScope.ServiceProvider.GetRequiredService<IOutboxStore>();
+        var runner = serviceScope.ServiceProvider.GetRequiredService<EfOrderSagaTransitionRunner>();
         var now = _timeProvider.GetUtcNow().UtcDateTime;
+
+        var compensationCandidates = new List<(Guid SagaId, OrderSagaStep Origin, Event Trigger)>();
 
         await sagaStore.ExecuteAsync(async () =>
         {
-            var commands = new List<Event>();
             var overdueSagas = await sagaStore.GetOverdueOrderSagas(OrderSagaType, now, stoppingToken);
 
             foreach (var saga in overdueSagas)
@@ -78,7 +81,13 @@ internal sealed partial class SagaReaperService : BackgroundService
 
                 if (saga.RetryCount >= _maxRetries)
                 {
-                    commands.AddRange(BeginCompensation(saga, currentStep, now));
+                    var trigger = new Event
+                    {
+                        CorrelationId = saga.CorrelationId,
+                        SagaId = saga.SagaId
+                    };
+                    var origin = OrderSagaStateMachine.GetLastCompletedStep(currentStep);
+                    compensationCandidates.Add((saga.SagaId, origin, trigger));
                     continue;
                 }
 
@@ -105,58 +114,13 @@ internal sealed partial class SagaReaperService : BackgroundService
             }
 
             await sagaStore.SaveChangesAsync(stoppingToken);
-            return commands;
-        });
-    }
-
-    private IReadOnlyList<Event> BeginCompensation(SagaInstance saga, OrderSagaStep currentStep, DateTime now)
-    {
-        var trigger = new Event
-        {
-            CorrelationId = saga.CorrelationId,
-            SagaId = saga.SagaId
-        };
-        var origin = OrderSagaStateMachine.GetLastCompletedStep(currentStep);
-        var snapshot = new OrderSagaStateSnapshot(
-            saga.SagaId,
-            saga.OrderSagaState!.OrderId,
-            currentStep,
-            saga.Status,
-            saga.OrderSagaState.LastStepResult,
-            saga.OrderSagaState.Amount,
-            ParseStep(saga.OrderSagaState.CompensationOrigin));
-        var result = OrderSagaStateMachine.BeginCompensation(snapshot, origin, trigger);
-
-        if (!result.Changed)
-        {
             return [];
-        }
-
-        saga.CurrentStep = result.State.CurrentStep.ToString();
-        saga.Status = result.State.Status;
-        saga.UpdatedAt = now;
-        saga.OrderSagaState.LastStepResult = result.State.LastStepResult;
-        saga.OrderSagaState.Amount = result.State.Amount;
-        saga.OrderSagaState.CompensationOrigin = result.State.CompensationOrigin?.ToString();
-        saga.LastCommandId = result.Commands.Count == 0 ? null : result.Commands[0].Id;
-        _timeoutScheduler.Apply(saga, now);
-        saga.Transitions.Add(new SagaTransition
-        {
-            SagaId = saga.SagaId,
-            FromStep = currentStep.ToString(),
-            ToStep = result.State.CurrentStep.ToString(),
-            Timestamp = now,
-            TriggerMessageId = trigger.Id,
-            TriggerKind = SagaTriggerKind.Timeout,
-            Error = "Saga step exceeded max retries; compensation started."
         });
 
-        SagaTelemetry.Compensation.Add(1,
-            new KeyValuePair<string, object?>("type", saga.SagaType));
-
-        LogSagaCompensating(_logger, saga.SagaId, saga.SagaType, currentStep.ToString(), origin.ToString());
-
-        return result.Commands;
+        foreach (var (sagaId, origin, trigger) in compensationCandidates)
+        {
+            await runner.BeginCompensation(sagaId, origin, trigger, stoppingToken);
+        }
     }
 
     private void MarkFailed(SagaInstance saga, DateTime now, string error)
@@ -185,9 +149,6 @@ internal sealed partial class SagaReaperService : BackgroundService
         LogSagaTimeoutFailure(_logger, saga.SagaId, saga.SagaType, fromStep, error);
     }
 
-    private static OrderSagaStep? ParseStep(string? value) =>
-        Enum.TryParse<OrderSagaStep>(value, out var parsed) ? parsed : null;
-
     [LoggerMessage(
         Level = LogLevel.Warning,
         Message = "Saga {SagaId} ({SagaType}) step {Step} is overdue at retry {RetryCount}/{MaxRetries}")]
@@ -198,16 +159,6 @@ internal sealed partial class SagaReaperService : BackgroundService
         string step,
         int retryCount,
         int maxRetries);
-
-    [LoggerMessage(
-        Level = LogLevel.Warning,
-        Message = "Saga {SagaId} ({SagaType}) step {Step} exceeded retries; compensating from {Origin}")]
-    private static partial void LogSagaCompensating(
-        ILogger logger,
-        Guid sagaId,
-        string sagaType,
-        string step,
-        string origin);
 
     [LoggerMessage(
         Level = LogLevel.Error,
